don't override -actionForLayer:forKey
I hadn’t thought this through properly before, but it makes sense now.
Super’s implementation (a NSView) returns the correct defaults for
actions such as “contents” or “position”. Additionally, it knows that
it needs to enable these implicit animations in an implicit animation
block.

So if we set the layer’s delegate to ourselves, we effectively have the
same behavior as a regular layer-backed view. That’s how it should have
been in the first place.